Move all paint icon methodes in IconPainter class
