+++ src/component/mxgraph/shape/task-shapes.ts
@@ -16,8 +16,8 @@
 
 import { MxGraphFactoryService } from '../../../service/MxGraphFactoryService';
 import { mxgraph } from 'ts-mxgraph';
-import { StyleConstant } from '../StyleConfigurator';
-import MxScaleFactorCanvas, { MxCanvasUtil } from '../extension/MxScaleFactorCanvas';
+import { StyleConstant } from '../StyleUtils';
+import IconPainter, { PaintParameter } from './IconPainter';
 
 const mxRectangleShape: typeof mxgraph.mxRectangleShape = MxGraphFactoryService.getMxGraphProperty('mxRectangleShape');
 
@@ -33,10 +33,10 @@ abstract class BaseTaskShape extends mxRectangleShape {
 
   public paintForeground(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void {
     super.paintForeground(c, x, y, w, h);
-    this.paintTaskIcon(c, x, y, w, h);
+    this.paintTaskIcon({ c, x, y, w, h, style: this.style });
   }
 
-  protected abstract paintTaskIcon(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void;
+  protected abstract paintTaskIcon(paintParameter: PaintParameter): void;
 }
 
 export class TaskShape extends BaseTaskShape {
@@ -45,8 +45,9 @@ export class TaskShape extends BaseTaskShape {
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  protected paintTaskIcon(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void {
+  protected paintTaskIcon(paintParameter: PaintParameter): void {
     // No symbol for the BPMN Task
+    IconPainter.paintEmptyIcon();
   }
 }
 
@@ -55,115 +56,8 @@ export class ServiceTaskShape extends BaseTaskShape {
     super(bounds, fill, stroke, strokewidth);
   }
 
-  // this implementation is adapted from the draw.io BPMN 'Service Task' stencil
-  // https://github.com/jgraph/drawio/blob/9394fb0f1430d2c869865827b2bbef5639f63478/src/main/webapp/stencils/bpmn.xml#L898
-  protected paintTaskIcon(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void {
-    // icon coordinates fill a 100x100 rectangle (approximately: 90x90 + foreground translation)
-    const canvas = MxCanvasUtil.getConfiguredCanvas(c, w, h, 100);
-    MxCanvasUtil.translateToStartingIconPosition(c, x, y, w, h, 20);
-
-    // background
-    this.drawIconBackground(canvas);
-
-    // foreground
-    const foregroundTranslation = 14 * canvas.scaleFactor;
-    c.translate(foregroundTranslation, foregroundTranslation);
-    this.drawIconForeground(canvas);
-  }
-
-  private drawIconBackground(canvas: MxScaleFactorCanvas): void {
-    canvas.begin();
-    canvas.moveTo(2.06, 24.62);
-    canvas.lineTo(10.17, 30.95);
-    canvas.lineTo(9.29, 37.73);
-    canvas.lineTo(0, 41.42);
-    canvas.lineTo(2.95, 54.24);
-    canvas.lineTo(13.41, 52.92);
-    canvas.lineTo(17.39, 58.52);
-    canvas.lineTo(13.56, 67.66);
-    canvas.lineTo(24.47, 74.44);
-    canvas.lineTo(30.81, 66.33);
-    canvas.lineTo(37.88, 67.21);
-    canvas.lineTo(41.57, 76.5);
-    canvas.lineTo(54.24, 73.55);
-    canvas.lineTo(53.06, 62.94);
-    canvas.lineTo(58.52, 58.52);
-    canvas.lineTo(67.21, 63.09);
-    canvas.lineTo(74.58, 51.88);
-    canvas.lineTo(66.03, 45.25);
-    canvas.lineTo(66.92, 38.62);
-    canvas.lineTo(76.5, 34.93);
-    canvas.lineTo(73.7, 22.26);
-    canvas.lineTo(62.64, 23.44);
-    canvas.lineTo(58.81, 18.42);
-    canvas.lineTo(62.79, 8.7);
-    canvas.lineTo(51.74, 2.21);
-    canvas.lineTo(44.81, 10.47);
-    canvas.lineTo(38.03, 9.43);
-    canvas.lineTo(33.75, 0);
-    canvas.lineTo(21.52, 3.24);
-    canvas.lineTo(22.7, 13.56);
-    canvas.lineTo(18.13, 17.54);
-    canvas.lineTo(8.7, 13.56);
-    canvas.close();
-
-    const arcStartX = 24.8;
-    const arcStartY = 39;
-    this.drawInnerCircle(canvas, arcStartX, arcStartY);
-  }
-
-  private drawIconForeground(canvas: MxScaleFactorCanvas): void {
-    canvas.begin();
-    canvas.moveTo(16.46, 41.42);
-    canvas.lineTo(24.57, 47.75);
-    canvas.lineTo(23.69, 54.53);
-    canvas.lineTo(14.4, 58.22);
-    canvas.lineTo(17.35, 71.04);
-    canvas.lineTo(27.81, 69.72);
-    canvas.lineTo(31.79, 75.32);
-    canvas.lineTo(27.96, 84.46);
-    canvas.lineTo(38.87, 91.24);
-    canvas.lineTo(45.21, 83.13);
-    canvas.lineTo(52.28, 84.01);
-    canvas.lineTo(55.97, 93.3);
-    canvas.lineTo(68.64, 90.35);
-    canvas.lineTo(67.46, 79.74);
-    canvas.lineTo(72.92, 75.32);
-    canvas.lineTo(81.61, 79.89);
-    canvas.lineTo(88.98, 68.68);
-    canvas.lineTo(80.43, 62.05);
-    canvas.lineTo(81.32, 55.42);
-    canvas.lineTo(90.9, 51.73);
-    canvas.lineTo(88.1, 39.06);
-    canvas.lineTo(77.04, 40.24);
-    canvas.lineTo(73.21, 35.22);
-    canvas.lineTo(77.19, 25.5);
-    canvas.lineTo(66.14, 19.01);
-    canvas.lineTo(59.21, 27.27);
-    canvas.lineTo(52.43, 26.23);
-    canvas.lineTo(48.15, 16.8);
-    canvas.lineTo(35.92, 20.04);
-    canvas.lineTo(37.1, 30.36);
-    canvas.lineTo(32.53, 34.34);
-    canvas.lineTo(23.1, 30.36);
-    canvas.close();
-
-    const arcStartX = 39.2;
-    const arcStartY = 55.8;
-    this.drawInnerCircle(canvas, arcStartX, arcStartY);
-
-    // fill the inner circle to mask the background
-    canvas.begin();
-    this.drawInnerCircle(canvas, arcStartX, arcStartY);
-  }
-
-  private drawInnerCircle(canvas: MxScaleFactorCanvas, arcStartX: number, arcStartY: number): void {
-    const arcRay = 13.5;
-    canvas.moveTo(arcStartX, arcStartY);
-    canvas.arcTo(arcRay, arcRay, 0, 1, 1, arcStartX + 2 * arcRay, arcStartY);
-    canvas.arcTo(arcRay, arcRay, 0, 0, 1, arcStartX, arcStartY);
-    canvas.close();
-    canvas.fillAndStroke();
+  protected paintTaskIcon(paintParameter: PaintParameter): void {
+    IconPainter.paintGearIcon(paintParameter);
   }
 }
 
@@ -172,91 +66,8 @@ export class UserTaskShape extends BaseTaskShape {
     super(bounds, fill, stroke, strokewidth);
   }
 
-  // implementation adapted from https://www.flaticon.com/free-icon/employees_554768
-  // use https://github.com/process-analytics/mxgraph-svg2shape to generate the xml stencil and port it to code
-  protected paintTaskIcon(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void {
-    // generated icon h="239.68" w="143.61"
-    const generatedIconHeight = 239;
-    const canvas = MxCanvasUtil.getConfiguredCanvas(c, w, h, generatedIconHeight, this.stroke, 0);
-    MxCanvasUtil.translateToStartingIconPosition(c, x, y, w, h, 20);
-
-    canvas.begin();
-    canvas.moveTo(124.31, 150.29);
-    canvas.lineTo(99.66, 141.03);
-    canvas.arcTo(6.43, 6.43, 0, 0, 1, 95.51, 135.03);
-    canvas.lineTo(95.51, 130.66);
-    canvas.arcTo(47.75, 47.75, 0, 0, 0, 119.51, 89.25);
-    canvas.lineTo(119.51, 71.25);
-    canvas.arcTo(47.62, 47.62, 0, 0, 0, 101.18, 33.64);
-    canvas.arcTo(29.35, 29.35, 0, 0, 0, 101.52, 29.14);
-    canvas.arcTo(29.68, 29.68, 0, 0, 0, 42.17, 29.14);
-    canvas.arcTo(29.24, 29.24, 0, 0, 0, 42.53, 33.63);
-    canvas.arcTo(47.65, 47.65, 0, 0, 0, 24.14, 71.23);
-    canvas.lineTo(24.14, 89.23);
-    canvas.arcTo(47.7, 47.7, 0, 0, 0, 48.19, 130.63);
-    canvas.lineTo(48.19, 135.03);
-    canvas.arcTo(6.43, 6.43, 0, 0, 1, 44.03, 141.03);
-    canvas.lineTo(19.31, 150.29);
-    canvas.arcTo(29.81, 29.81, 0, 0, 0, 0.09, 178.03);
-    canvas.lineTo(0.09, 233.51);
-    canvas.arcTo(5.63, 5.63, 0, 1, 0, 11.34, 233.51);
-    canvas.lineTo(11.34, 178.03);
-    canvas.arcTo(18.19, 18.19, 0, 0, 1, 11.57, 175.17);
-    canvas.lineTo(20.5, 184.11);
-    canvas.arcTo(12.32, 12.32, 0, 0, 1, 24.14, 192.89);
-    canvas.lineTo(24.14, 233.51);
-    canvas.arcTo(5.63, 5.63, 0, 1, 0, 35.39, 233.51);
-    canvas.lineTo(35.39, 192.93);
-    canvas.arcTo(23.5, 23.5, 0, 0, 0, 28.46, 176.2);
-    canvas.lineTo(17.04, 164.78);
-    canvas.arcTo(18.34, 18.34, 0, 0, 1, 23.29, 160.78);
-    canvas.lineTo(43.65, 153.15);
-    canvas.lineTo(66.22, 175.72);
-    canvas.lineTo(66.22, 233.51);
-    canvas.arcTo(5.63, 5.63, 0, 1, 0, 77.47, 233.51);
-    canvas.lineTo(77.47, 175.76);
-    canvas.lineTo(100.04, 153.19);
-    canvas.lineTo(120.4, 160.82);
-    canvas.arcTo(18.39, 18.39, 0, 0, 1, 126.65, 164.82);
-    canvas.lineTo(115.24, 176.24);
-    canvas.arcTo(23.5, 23.5, 0, 0, 0, 108.31, 192.93);
-    canvas.lineTo(108.31, 233.55);
-    canvas.arcTo(5.63, 5.63, 0, 1, 0, 119.56, 233.55);
-    canvas.lineTo(119.56, 192.93);
-    canvas.arcTo(12.35, 12.35, 0, 0, 1, 123.19, 184.15);
-    canvas.lineTo(132.13, 175.22);
-    canvas.arcTo(18, 18, 0, 0, 1, 132.36, 178.08);
-    canvas.lineTo(132.36, 233.56);
-    canvas.arcTo(5.63, 5.63, 0, 0, 0, 143.61, 233.56);
-    canvas.lineTo(143.61, 178.03);
-    canvas.arcTo(29.81, 29.81, 0, 0, 0, 124.31, 150.29);
-    canvas.close();
-    canvas.moveTo(71.85, 10.72);
-    canvas.arcTo(18.46, 18.46, 0, 0, 1, 90.17, 27.18);
-    canvas.arcTo(47.68, 47.68, 0, 0, 0, 53.53, 27.18);
-    canvas.arcTo(18.44, 18.44, 0, 0, 1, 71.85, 10.72);
-    canvas.close();
-    canvas.moveTo(35.39, 71.23);
-    canvas.arcTo(36.46, 36.46, 0, 0, 1, 108.31, 71.23);
-    canvas.lineTo(108.31, 77.4);
-    canvas.curveTo(82.12, 75.4, 56.97, 60.55, 56.71, 60.4);
-    canvas.arcTo(5.62, 5.62, 0, 0, 0, 48.78, 62.71);
-    canvas.curveTo(46.24, 67.79, 40.45, 71.89, 35.39, 74.62);
-    canvas.close();
-    canvas.moveTo(35.39, 89.23);
-    canvas.lineTo(35.39, 87.08);
-    canvas.curveTo(40.55, 84.85, 49.73, 80.08, 55.67, 72.66);
-    canvas.curveTo(64.83, 77.46, 85.92, 87.21, 108.31, 88.66);
-    canvas.lineTo(108.31, 89.24);
-    canvas.arcTo(36.46, 36.46, 0, 1, 1, 35.39, 89.24);
-    canvas.close();
-    canvas.moveTo(71.85, 165.45);
-    canvas.lineTo(54.06, 147.69);
-    canvas.arcTo(17.7, 17.7, 0, 0, 0, 59.43, 135.32);
-    canvas.arcTo(47.57, 47.57, 0, 0, 0, 84.27, 135.32);
-    canvas.arcTo(17.7, 17.7, 0, 0, 0, 89.64, 147.69);
-    canvas.close();
-    canvas.fill();
+  protected paintTaskIcon(paintParameter: PaintParameter): void {
+    IconPainter.paintWomanIcon(paintParameter);
   }
 }
 
@@ -266,6 +77,6 @@ export class ReceiveTaskShape extends BaseTaskShape {
     this.gradient = 'Salmon';
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-empty-function,@typescript-eslint/no-unused-vars
-  protected paintTaskIcon(c: mxgraph.mxXmlCanvas2D, x: number, y: number, w: number, h: number): void {}
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-function
+  protected paintTaskIcon(paintParameter: PaintParameter): void {}
 }
